Устранено неравенство температур
Устранено неравенство температур с датчиков DS18B20 и BME280 при одинаковых условиях. Проблема была в неявном преобразовании типов

+++ main.ino
@@ -26,7 +26,7 @@
     byte ext_temp_data[BYTE_ITERATOR];
     byte HighByte;
     byte LowByte; 
-    double external_temperature;
+    float external_temperature;
     byte tempInt; 
     byte tempFloat;
     Serial_measure _temp_ext;
